requirements.txt
psycopg2-binary
pandas
python-dotenv
requests
black
pylint
aiohttp
asyncpg
pyyaml
minio
loguru
tqdm
boto3
sqlalchemy[asyncio]